Feature: 9 van subscribe - Refactor part 1 - TODO: write tests for `sendMinimalConsensusRange` `handleMinimalConsensusSubscription`
